+++ server/component/cms/CmsView.php
@@ -1063,14 +1063,14 @@ class CmsView extends BaseView
                 "is_required" => 1,
                 "items" => $this->model->get_db()->fetch_table_as_select_values('lookups', 'id', array('lookup_value'), 'WHERE type_code=:tcode', array(":tcode" => pageAccessTypes))
             ));
-        } else if ($field['type'] == "select-formName") {
+        } else if ($field['type'] == "select-data_table") {
             $children[] = new BaseStyleComponent("select", array(
                 "value" => $field['content'],
                 "name" => $field_name_prefix . "[content]",
                 "max" => 10,
                 "live_search" => 1,
                 "is_required" => 1,
-                "items" => $this->model->get_db()->fetch_table_as_select_values('dataTables', 'id', array('name', 'displayName'), '')
+                "items" => $this->model->get_db()->fetch_table_as_select_values('view_dataTables', 'id', array('name'), '')
             ));
         } else if ($field['type'] == "select-plugin") {
             $children[] = new BaseStyleComponent("select", array(
@@ -1186,12 +1186,12 @@ class CmsView extends BaseView
                 "items" => $this->model->get_db()->fetch_table_as_select_values('lookups', 'id', array('lookup_value'), 'WHERE type_code=:tcode', array(":tcode" => pageAccessTypes))
             ));
         }
-        else if ($field['type'] == "select-formName") {
+        else if ($field['type'] == "select-data_table") {
             $children[] = new BaseStyleComponent("select", array(
                 "value" => $field['content'],
                 "name" => $field['name'],
                 "disabled" => 1,
-                "items" => $this->model->get_db()->fetch_table_as_select_values('dataTables', 'id', array('name', 'displayName'), '')
+                "items" => $this->model->get_db()->fetch_table_as_select_values('view_dataTables', 'id', array('name'), '')
             ));
         }
         else if($field['type'] == "select-plugin")