+++ server/db/update_scripts/31_update_v6.16.0_v7.0.0.sql
@@ -608,6 +608,9 @@ SET `name` = 'data_table'
 WHERE `name` = 'formName';
 
 -- rename fieldType `select-formName` to `select-data_table`
+UPDATE fieldType
+SET `name` = 'select-data_table'
+WHERE `name` = 'select-formName';