+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - add `scope` to `dataConfig`, now for each data source could be added a separated scope. If a scope is added then the variables will be set as `scope.var_name` and later could be interpolated like this
  - add field `formName` to style `showUserInput`; Set the data from `source` field to the formName and remove the `source` field
  - rename style field `formName` to `data_table`
+ - rename fieldType `select-formName` to `select-data_table`
 
 # v6.16.4
  - change function `get_user_group` in `Condition` to be a public function